Make `require` use a module
Now have a module that exports the `require` & `define` functions.
